feat: Eventless component (#29)
- update JSDoc comments 1479058
- change preset to base 0dad6c3
- use more minify plugins 03c4437
- feat: add eventless component option 10e7959
