+++ .eslintrc.json
@@ -409,7 +409,10 @@
 
                 "@typescript-eslint/type-annotation-spacing": ["error", { "before": false, "after": true }],
 
-                "@typescript-eslint/unified-signatures": "error"
+                "@typescript-eslint/unified-signatures": "error",
+
+                "require-jsdoc": "error",
+                "valid-jsdoc": "off"
             }
         },
         {
