+++ package.json
@@ -15,14 +15,16 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.10.4",
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@types/babel__core": "^7.1.9",
     "@types/node": "^14.0.20",
     "@typescript-eslint/eslint-plugin": "^3.6.0",
     "@typescript-eslint/parser": "^3.6.0",
     "babel-plugin-loop-optimizer": "^1.4.1",
-    "babel-plugin-minify-constant-folding": "^0.5.0",
-    "babel-plugin-minify-dead-code-elimination": "^0.5.1",
     "babel-plugin-transform-merge-sibling-variables": "^6.9.4",
     "babel-plugin-transform-minify-booleans": "^6.9.4",
+    "babel-preset-minify": "^0.5.1",
     "eslint": "^7.4.0",
     "eslint-config-recommended": "^4.1.0",
     "eslint-plugin-prefer-arrow": "^1.2.1",
